refactor(#151): dedup the two WASM adapters into adapter_common (re-land)
Re-land of #151 (PR #167), stranded when it merged into the fix/141 base
branch after that branch squash-landed in main. Recovered onto current main.

Hoists the byte-identical helpers (rows_to_maps, build_pk_text_expr,
row_maps_to_metadata, content_hash, generate_batch_sql, parse_table_info) from
both adapters into adapter_common; both call adapter_common::<fn>. Pure move,
hashes/SQL unchanged. Also the 10 needless_borrow lib.rs fixes for the
wasm-target clippy gate. cached_table_info stays per-adapter.
