+++ crates/smugglr-wasm/src/lib.rs
@@ -16,6 +16,7 @@
 
 #![cfg(target_arch = "wasm32")]
 
+mod adapter_common;
 mod fetch_adapter;
 mod local_adapter;
 
@@ -785,7 +786,7 @@ impl Smugglr {
             )
         })?;
         let dry_run = dry_run.unwrap_or(false);
-        let tables = get_sync_tables(&self.source, &dest, &self.sync_config).await?;
+        let tables = get_sync_tables(&self.source, dest, &self.sync_config).await?;
         let conflict = self.sync_config.conflict_resolution;
         let batch_size = self.sync_config.batch_size;
 
@@ -793,7 +794,7 @@ impl Smugglr {
         for table in &tables {
             let diff = diff_table_cached(
                 &self.source,
-                &dest,
+                dest,
                 &self.source_cache,
                 &self.dest_cache,
                 table,
@@ -808,7 +809,7 @@ impl Smugglr {
             } else {
                 transfer_rows(
                     &self.source,
-                    &dest,
+                    dest,
                     table,
                     &to_push,
                     batch_size,
@@ -844,7 +845,7 @@ impl Smugglr {
             )
         })?;
         let dry_run = dry_run.unwrap_or(false);
-        let tables = get_sync_tables(&self.source, &dest, &self.sync_config).await?;
+        let tables = get_sync_tables(&self.source, dest, &self.sync_config).await?;
         let conflict = self.sync_config.conflict_resolution;
         let batch_size = self.sync_config.batch_size;
 
@@ -852,7 +853,7 @@ impl Smugglr {
         for table in &tables {
             let diff = diff_table_cached(
                 &self.source,
-                &dest,
+                dest,
                 &self.source_cache,
                 &self.dest_cache,
                 table,
@@ -866,7 +867,7 @@ impl Smugglr {
                 to_pull.len()
             } else {
                 let n = transfer_rows(
-                    &dest,
+                    dest,
                     &self.source,
                     table,
                     &to_pull,
@@ -907,7 +908,7 @@ impl Smugglr {
             )
         })?;
         let dry_run = dry_run.unwrap_or(false);
-        let tables = get_sync_tables(&self.source, &dest, &self.sync_config).await?;
+        let tables = get_sync_tables(&self.source, dest, &self.sync_config).await?;
         let conflict = self.sync_config.conflict_resolution;
         let batch_size = self.sync_config.batch_size;
 
@@ -915,7 +916,7 @@ impl Smugglr {
         for table in &tables {
             let diff = diff_table_cached(
                 &self.source,
-                &dest,
+                dest,
                 &self.source_cache,
                 &self.dest_cache,
                 table,
@@ -932,7 +933,7 @@ impl Smugglr {
             } else {
                 transfer_rows(
                     &self.source,
-                    &dest,
+                    dest,
                     table,
                     &to_push,
                     batch_size,
@@ -945,7 +946,7 @@ impl Smugglr {
                 to_pull.len()
             } else {
                 let n = transfer_rows(
-                    &dest,
+                    dest,
                     &self.source,
                     table,
                     &to_pull,
